==> 06_Model_Binary_Logistic.ipynb <==
#import
import numpy as np
import pandas as pd
pd.set_option('display.float_format', lambda x: '%.3f' % x) #Set the pandas display format to suppress scientific notation
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.metrics import confusion_matrix, accuracy_score, recall_score
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split

#import data
df = pd.read_csv("data/binary_state_kickstarter_projects.csv")
df.head()

# Scaling nummerical data
numerical = df[['goal', 'duration', 'category_encoded', 'country_encoded']]
numerical2 = df[['backers', 'pledged', 'goal', 'duration']]
numerical3 = df[['goal', 'duration']]

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()

# Import encoder
from sklearn.preprocessing import LabelEncoder

#Defining X and y
X = scaler.fit_transform(numerical3)
y = df['state']

pd.DataFrame(X).describe()

# Splitting the dataset into the Training set and Test set
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 21)

# We import and train a Logistic Regression
from sklearn.linear_model import LogisticRegression
algorithm = LogisticRegression(random_state = 22)
algorithm.fit(X_train, y_train)

#Cross Validation
from sklearn.model_selection import cross_val_score

accuracies = cross_val_score(estimator = algorithm, X = X_train, y = y_train, cv = 5)

print(accuracies.mean())
print(accuracies.std())

print(algorithm.score(X_test, y_test))

# Predicting the Test set results
y_pred = algorithm.predict(X_test)
# Making a classification report
from sklearn.metrics import classification_report

cm = classification_report(y_test, y_pred)
print(cm)

# Transforming nummerical labels to original state types
true_projects = y_test
predicted_projects = y_pred

# Creating a pandas DataFrame and cross-tabulation
df = pd.DataFrame({'true_projects': true_projects, 'predicted_projects': predicted_projects}) 
pd.crosstab(df.true_projects, df.predicted_projects)
