==> 10_Linear_Regression_Model.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score

import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv('data/cleaned_encoded_kickstarter_projects.csv')
df.head()

df.columns

df.info()

print(df.describe())

# Define feature and target variable
X = df[['goal', 'duration', 'launch_month', 'category_encoded', 'subcategory_encoded', 'country_encoded']]
y = df['pledged']

X.shape

# Apply logarithmic transformation to 'goal' and 'pledged' columns
X['goal'] = np.log1p(X['goal'])  # log1p is used to handle zero values
y = np.log1p(y)

y.describe()

# Fit linear regression model
lin_reg = LinearRegression()
lin_reg.fit(X, y)

# Calculate intercept and coefficient
intercept = lin_reg.intercept_
coefficients = lin_reg.coef_
print("Intercept:", intercept.round(5))
print("Coefficients:", coefficients.round(5))

df[['goal', 'duration', 'launch_month', 'category_encoded', 'subcategory_encoded', 'country_encoded']].corr()

# Calculate r-squared 
y_hat = lin_reg.predict(X)
print("R-squared:", r2_score(y, y_hat).round(3))

from sklearn.preprocessing import PolynomialFeatures

# Create PolynomialFeatures object with interaction_only=True
poly = PolynomialFeatures(degree=3, interaction_only=True, include_bias=False)

# Fit and transform the data
interaction_features = poly.fit_transform(X)

# Create a new df with interaction features
feature_names = poly.get_feature_names_out(input_features=X.columns)
interaction_df = pd.DataFrame(interaction_features, columns=feature_names)

# Split data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(interaction_df, y, test_size=0.25, random_state=42)

# Scale features
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

# Train the linear regression model
model = LinearRegression()
model.fit(X_train_scaled, y_train)

# Make predictions
y_pred = model.predict(X_test_scaled)

# Evaluate the model
mse = mean_squared_error(y_test, y_pred)
print(f'Mean Squared Error: {mse}')

# Inverse transform the predictions to the original scale
y_pred_original_scale = np.expm1(y_pred)  # Inverse of log1p is expm1
y_test_original_scale = np.expm1(y_test)

# Evaluating the model with original target values
mse_original_scale = mean_squared_error(y_test_original_scale, y_pred_original_scale)
print(f'Mean Squared Error (Original Scale): {mse_original_scale}')
